backend/requirements.txt
flask
torch
transformers
torchvision
Pillow
scikit-learn
numpy
